requirements.txt
allure-pytest==2.10.0
allure-python-commons==2.10.0
async-generator==1.10
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
colorama==0.4.5
execnet==1.9.0
Faker==14.1.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
loguru==0.6.0
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-forked==1.4.0
pytest-json-report==1.5.0
pytest-metadata==2.0.2
pytest-ordering==0.6
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
python-dateutil==2.8.2
PyYAML==6.0
selenium==4.4.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
win32-setctime==1.1.0
wsproto==1.2.0
